requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
Brotli==1.1.0
celery==5.4.0
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
chardet==5.2.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==44.0.1
cssselect2==0.7.0
daphne==4.1.2
dj-database-url==2.3.0
Django==5.1.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.9
Faker==36.1.1
fonttools==4.56.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
kombu==5.4.2
msgpack==1.1.0
packaging==24.2
pillow==11.1.0
prompt_toolkit==3.0.50
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydyf==0.11.0
PyJWT==2.9.0
pyOpenSSL==25.0.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
reportlab==4.3.1
service-identity==24.2.0
setuptools==75.8.2
six==1.17.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
weasyprint==64.1
webencodings==0.5.1
zope.interface==7.2
zopfli==0.2.3.post1
